Fixed a lot of minor warnings

+++ src/tightdb/alloc_slab.cpp
@@ -140,9 +140,9 @@ MemRef SlabAlloc::Alloc(size_t size)
 
     // Else, allocate new slab
     const size_t multible = 256 * ((size / 256) + 1); // FIXME: Not an english word. Also, is this the intended rounding behavior?
-    const size_t slabsBack = m_slabs.is_empty() ? m_baseline : m_slabs.back().offset;
+    const size_t slabsBack = m_slabs.is_empty() ? m_baseline : size_t(m_slabs.back().offset);
     const size_t doubleLast = m_slabs.is_empty() ? 0 :
-        (slabsBack - ((m_slabs.size() == 1) ? size_t(0) : m_slabs.back(-2).offset)) * 2;
+        (slabsBack - ((m_slabs.size() == 1) ? size_t(0) : size_t(m_slabs.back(-2).offset)) * 2);
     const size_t newsize = multible > doubleLast ? multible : doubleLast;
 
     // Allocate memory
@@ -259,7 +259,7 @@ void* SlabAlloc::Translate(size_t ref) const
         const size_t ndx = m_slabs.column().offset.find_pos(ref);
         TIGHTDB_ASSERT(ndx != not_found);
 
-        const size_t offset = ndx ? m_slabs[ndx-1].offset : m_baseline;
+        const size_t offset = ndx ? size_t(m_slabs[ndx-1].offset) : m_baseline;
         return (char*)(intptr_t)m_slabs[ndx].pointer + (ref - offset);
     }
 }
